Refactoring wire-runtime to proper Hierarchical Multi-Platform Project.
