+++ wire-runtime/build.gradle.kts
@@ -42,6 +42,9 @@ kotlin {
   }
 
   if (System.getProperty("kwasm", "true").toBoolean()) {
+    wasmJs {
+      browser()
+    }
     @OptIn(ExperimentalWasmDsl::class)
     wasmWasi {
       nodejs()
@@ -71,15 +74,46 @@ kotlin {
         implementation(libs.kotlin.test.junit)
       }
     }
+    val nonJvmMain by creating {
+      dependsOn(commonMain)
+    }
+    val nonJvmTest by creating {
+      dependsOn(commonTest)
+    }
+    if (System.getProperty("knative", "true").toBoolean()) {
+      val nativeMain by getting {
+        dependsOn(nonJvmMain)
+      }
+      val nativeTest by getting {
+        dependsOn(nonJvmTest)
+      }
+    }
     if (System.getProperty("kjs", "true").toBoolean()) {
+      val jsMain by getting {
+        dependsOn(nonJvmMain)
+      }
       val jsTest by getting {
+        dependsOn(nonJvmTest)
         dependencies {
           implementation(libs.kotlin.test.js)
         }
       }
     }
     if (System.getProperty("kwasm", "true").toBoolean()) {
+      val wasmJsMain by getting {
+        dependsOn(nonJvmMain)
+      }
+      val wasmJsTest by getting {
+        dependsOn(nonJvmTest)
+        dependencies {
+          implementation(libs.kotlin.test.wasm.js)
+        }
+      }
+      val wasmWasiMain by getting {
+        dependsOn(nonJvmMain)
+      }
       val wasmWasiTest by getting {
+        dependsOn(nonJvmTest)
         dependencies {
           implementation(libs.kotlin.test)
         }
