+++ wire-runtime/src/nonJvmMain/kotlin/com/squareup/wire/Duration.kt
