+++ wire-runtime/src/nonJvmMain/kotlin/com/squareup/wire/EnumAdapter.kt
