+++ wire-runtime/src/nonJvmMain/kotlin/com/squareup/wire/MessageSink.kt
