+++ wire-runtime/src/nonJvmMain/kotlin/com/squareup/wire/MessageSource.kt
