+++ wire-runtime/src/wasmJsMain/kotlin/com/squareup/wire/internal/-Platform.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Square, Inc.
+ * Copyright (C) 2025 Square, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.wire
+package com.squareup.wire.internal
 
-actual interface MessageSink<in T : Any> {
-  actual fun write(message: T)
-
-  actual fun cancel()
-
-  actual fun close()
-}
+actual typealias JsName = kotlin.js.JsName
